DIO-AWS/src/main/java/org/example/Pessoa.java
package org.example;

import java.time.temporal.ChronoUnit;
import java.time.LocalDateTime;

public class Pessoa {

    private String nome;
    private LocalDateTime nascimento;

    public Pessoa(String nome, LocalDateTime nascimento) {
        this.nome = nome;
        this.nascimento = nascimento;
    }

    public int getIDade(){
        return (int) ChronoUnit.YEARS.between(nascimento, LocalDateTime.now());
    }

    public boolean ehMaiorDeIdade(){
        return getIDade() >= 18;
    }
}
